8289910: unify os::message_box across posix platforms
Backport-of: e5491a2605177a9dca87a060d99aa5ea4fd4a239
